Change HasGlobalRole customizations into HasGlobalRole queries
