+++ service/src/main/kotlin/fi/espoo/evaka/shared/security/actionrule/HasGlobalRole.kt
@@ -4,10 +4,17 @@
 
 package fi.espoo.evaka.shared.security.actionrule
 
+import fi.espoo.evaka.shared.AttachmentId
+import fi.espoo.evaka.shared.PersonId
 import fi.espoo.evaka.shared.auth.AuthenticatedUser
 import fi.espoo.evaka.shared.auth.UserRole
+import fi.espoo.evaka.shared.db.Database
+import fi.espoo.evaka.shared.security.AccessControlDecision
 import fi.espoo.evaka.shared.utils.toEnumSet
+import org.jdbi.v3.core.kotlin.mapTo
 import java.util.EnumSet
+
+private typealias Filter<T> = (tx: Database.Read, user: AuthenticatedUser.Employee, targets: Set<T>) -> Iterable<T>
 
 data class HasGlobalRole(val oneOf: EnumSet<UserRole>) : StaticActionRule, ActionRuleParams<HasGlobalRole> {
     init {
@@ -20,5 +27,49 @@ data class HasGlobalRole(val oneOf: EnumSet<UserRole>) : StaticActionRule, Actio
 
     override fun merge(other: HasGlobalRole): HasGlobalRole = HasGlobalRole(
         (this.oneOf.asSequence() + other.oneOf.asSequence()).toEnumSet()
+    )
+
+    private class Query<T>(private val filter: Filter<T>) : DatabaseActionRule.Query<T, HasGlobalRole> {
+        override fun execute(
+            tx: Database.Read,
+            user: AuthenticatedUser,
+            targets: Set<T>
+        ): Map<T, DatabaseActionRule.Deferred<HasGlobalRole>> = when (user) {
+            is AuthenticatedUser.Employee -> filter(tx, user, targets).associateWith { Deferred(user.globalRoles) }
+            else -> emptyMap()
+        }
+    }
+    private class Deferred(private val globalRoles: Set<UserRole>) : DatabaseActionRule.Deferred<HasGlobalRole> {
+        override fun evaluate(params: HasGlobalRole): AccessControlDecision = if (globalRoles.any { params.oneOf.contains(it) }) {
+            AccessControlDecision.Permitted(params)
+        } else {
+            AccessControlDecision.None
+        }
+    }
+
+    fun andAttachmentWasUploadedByAnyEmployee() = DatabaseActionRule(
+        this,
+        Query<AttachmentId> { tx, _, ids ->
+            tx.createQuery(
+                """
+SELECT attachment.id
+FROM attachment
+JOIN evaka_user ON uploaded_by = evaka_user.id
+WHERE attachment.id = ANY(:ids)
+AND evaka_user.type = 'EMPLOYEE'
+                """.trimIndent()
+            )
+                .bind("ids", ids.toTypedArray())
+                .mapTo()
+        }
+    )
+
+    fun andPersonHasSsnAddingEnabled() = DatabaseActionRule(
+        this,
+        Query<PersonId> { tx, _, ids ->
+            tx.createQuery("SELECT id FROM person WHERE id = ANY(:ids) AND ssn_adding_disabled IS FALSE")
+                .bind("ids", ids.toTypedArray())
+                .mapTo()
+        }
     )
 }
